+++ R/zzz.R
@@ -40,8 +40,10 @@ on_package_load <- function(pkg, expr){
   }
 }
 .onAttach <- function(...){
-  options("cheapr.cores" = getOption("cheapr.cores", 1))
+  options("cheapr.cores" = getOption("cheapr.cores", 1),
+          "cheapr.digits" = getOption("cheapr.digits", 2))
 }
 .onUnload <- function(libname, pkgname){
-  options(cheapr.cores = NULL)
+  options(cheapr.cores = NULL,
+          cheapr.digits = NULL)
 }