+++ NEWS.md
@@ -1,5 +1,9 @@
 # cheapr (Development version)
 
+* `overview` now always returns an object of class "overview". It also returns
+the number of observations instead of rows so that it makes sense 
+for vector summaries as well as data frame summaries.
+
 * `sequence_` has been optimised and rewritten in C++. It now only checks for
 integer overflow when both `from` and `by` are integer vectors.